Laba7

+++ variant20/app.py
@@ -1,24 +1,66 @@
-from flask import Flask, request, jsonify
-from marshmallow import ValidationError
-from variant20.schemas import UserSchema, AdvertisementSchema
-from variant20.database import Session, User, Advertisement, ModifierEnum
 import bcrypt
+from flask import Flask, request, jsonify
+from flask_bcrypt import Bcrypt
+from flask_httpauth import HTTPBasicAuth
+from marshmallow import ValidationError
 
+from variant20.database import Session, User, Advertisement, RoleEnum
+from variant20.schemas import UserSchema, AdvertisementSchema
 
 app = Flask(__name__)
 session = Session()
+auth = HTTPBasicAuth()
+bcryptt = Bcrypt(app)
+
+
+@auth.verify_password
+def verify_password(email, password):
+    user = session.query(User).filter(User.email == email).first()
+    if user is not None and \
+            bcrypt.checkpw(password.encode('utf8'), user.password_hash.encode('utf8')):
+        return user
+    else:
+        return "Unauthorized"
+
+
+@app.route('/', methods=['GET'])
+@auth.login_required()
+def g():
+    user = auth.current_user()
+    if user == "Unauthorized":
+        return "You do not have sufficient editing rights", 403
+    return "ou may"
+
+
+@app.route('/a')
+@auth.login_required
+def index():
+    user = auth.current_user()
+    if user == "Unauthorized":
+        return "UNAUTH"
+    return "Hello, {}!".format(user.email)
 
 
 @app.route('/advertisements', methods=['GET'])
+@auth.login_required
 def get_ads():
-    ads = session.query(Advertisement).all()
+    user: object = auth.current_user()
+    if user == "Unauthorized":
+        ads = session.query(Advertisement).filter(Advertisement.modifier == "public").all()
+    else:
+        ads = session.query(Advertisement).all()
     return jsonify(AdvertisementSchema(many=True).dump(ads)), 200
 
 
 @app.route('/advertisements', methods=['POST'])
+@auth.login_required
 def create_ad():
+    user = auth.current_user()
+    if user == "Unauthorized":
+        return "You do not have sufficient editing rights", 403
     ad_schema = AdvertisementSchema()
     data = request.get_json()
+    data['user_id'] = auth.current_user().id
     try:
         ad = ad_schema.load(data)
     except ValidationError as err:
@@ -29,13 +71,21 @@ def create_ad():
 
 
 @app.route('/advertisements/<ad_topic>', methods=['GET'])
+@auth.login_required
 def get_ads_by_topic(ad_topic):
     ad_schema = AdvertisementSchema(only=['topic'])
     try:
         ad_schema.load({'topic': ad_topic})
     except ValidationError as err:
         return err.messages, 400
-    ads = session.query(Advertisement).filter_by(topic=ad_topic).all()
+    user: object = auth.current_user()
+    if user == "Unauthorized":
+        ads = session.query(Advertisement).filter(
+            Advertisement.modifier == "public",
+            Advertisement.topic == ad_topic
+            ).all()
+    else:
+        ads = session.query(Advertisement).filter(Advertisement.topic == ad_topic).all()
     return jsonify(AdvertisementSchema(many=True).dump(ads)), 200
 
 
@@ -49,47 +99,63 @@ def get_ad(ad_topic, ad_id):
     ad = session.query(Advertisement).filter(Advertisement.id == ad_id, Advertisement.topic == ad_topic).first()
     if ad is None:
         return 'Advertisement not found', 404
+    app.logger.info(ad.modifier)
     return jsonify(AdvertisementSchema().dump(ad)), 200
 
 
 @app.route('/advertisements/<ad_topic>/<ad_id>', methods=['PUT'])
+@auth.login_required
 def edit_ad(ad_topic, ad_id):
+    user = auth.current_user()
+    if user == "Unauthorized":
+        return "You do not have sufficient editing rights", 403
     ad_schema = AdvertisementSchema(only=['id', 'topic'])
     try:
         ad_schema.load({'id': ad_id, 'topic': ad_topic})
     except ValidationError as err:
         return err.messages, 400
     ad = session.query(Advertisement).filter(Advertisement.id == ad_id, Advertisement.topic == ad_topic).first()
+
     if ad is None:
         return 'Advertisement not found', 404
     ad_schema = AdvertisementSchema()
     data = request.get_json()
-    try:
-        ad_schema.load(data)
-    except ValidationError as err:
-        return err.messages, 400
-    ad.summary = data['summary']
-    ad.description = data['description']
-    ad.topic = data['topic']
-    ad.modifier = data['modifier']
-    ad.user_id = data['user_id']
-    session.commit()
-    return ad_schema.dump(ad), 201
+    if auth.current_user().role == RoleEnum.master or auth.current_user().id == ad.user_id == auth.current_user().id:
+        try:
+            ad_schema.load(data)
+        except ValidationError as err:
+            return err.messages, 400
+        ad.summary = ad.summary if 'summary' not in data else data['summary']
+        ad.description = ad.description if 'description' not in data else data['description']
+        ad.topic = ad.topic if 'topic' not in data else data['topic']
+        ad.modifier = ad.modifier if 'modifier' not in data else data['modifier']
+        ad.user_id = ad.user_id if 'user_id' not in data else data['user_id']
+        session.commit()
+        return ad_schema.dump(ad), 201
+    return "You do not have sufficient editing rights", 403
 
 
 @app.route('/advertisements/<ad_topic>/<ad_id>', methods=['DELETE'])
+@auth.login_required
 def delete_ad(ad_topic, ad_id):
+    user = auth.current_user()
+    if user == "Unauthorized":
+        return "You do not have sufficient editing rights", 403
     ad_schema = AdvertisementSchema(only=['id', 'topic'])
     try:
         ad_schema.load({'id': ad_id, 'topic': ad_topic})
     except ValidationError as err:
         return err.messages, 400
     ad = session.query(Advertisement).filter(Advertisement.id == ad_id, Advertisement.topic == ad_topic).first()
-    if ad is None:
-        return 'Advertisement not found', 404
-    session.delete(ad)
-    session.commit()
-    return "Advertisement deleted successfully"
+    if auth.current_user().role == RoleEnum.master or auth.current_user().id == ad.user_id == auth.current_user().id:
+        ad = session.query(Advertisement).filter(Advertisement.id == ad_id, Advertisement.topic == ad_topic).first()
+        if ad is None:
+            return 'Advertisement not found', 404
+        session.delete(ad)
+        session.commit()
+
+        return "Advertisement deleted successfully", 200
+    return "You do not have sufficient editing rights", 403
 
 
 @app.route('/users', methods=['GET'])
@@ -102,13 +168,14 @@ def get_users():
 def create_user():
     user_schema = UserSchema()
     data = request.get_json()
-    if not session.query(User).filter(User.name == data['name']).first() is None:
-        return 'The name is already taken', 400
+
     if not session.query(User).filter(User.email == data['email']).first() is None:
         return 'The email is already taken', 400
     try:
         user = user_schema.load({'name': data['name'], 'email': data['email'],
-                                 'password_hash':bcrypt.hashpw(data['password_hash'].encode('utf8'), bcrypt.gensalt())})
+                                 'password_hash': bcrypt.hashpw(data['password_hash'].encode('utf8'),
+                                                                bcrypt.gensalt())})
+        user.role = user.role if not 'role' in data else data['role']
     except ValidationError as err:
         return err.messages, 400
     session.add(user)
@@ -129,25 +196,7 @@ def get_user(user_id):
     return jsonify(UserSchema().dump(user)), 200
 
 
-@app.route('/sign-in', methods=['GET'])
-def sign_in():
-    email = request.args.get('email')
-    password = request.args.get('password')
-
-    user = session.query(User).filter(User.email == email).first()
-    if user is None:
-        return "Email not found", 400
-
-    if not bcrypt.checkpw(password.encode('utf8'), user.password_hash.encode('utf8')):
-        return "Wrong password", 400
-
-    return jsonify(UserSchema().dump(user)), 200
-
-
-@app.route('/logout', methods=['GET'])
-def logout():
-    return 'Logout'
 
 
 if __name__ == '__main__':
-    app.run()
+    app.run(debug=True)
